Added supabase auth and initial handling of signup and login
